django/requirements.txt
amqp==5.3.1
anyio==4.7.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
base58==2.1.1
billiard==4.2.1
bip32==4.0
bip44==0.1.5
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coincurve==20.0.0
cron-descriptor==1.4.5
cryptography==44.0.0
cytoolz==1.0.1
decorator==5.1.1
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-encrypted-model-fields==0.6.5
django-timezone-field==7.0
eth-hash==0.7.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
executing==2.1.0
factory_boy==3.3.1
Faker==33.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
mnemonic==0.21
packaging==24.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
segno==1.6.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
toolz==1.0.0
traitlets==5.14.3
tronpy==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
